update tests for cli.py and added smoke tests
